aarch64 fix in mistralrs-quant
On aarch64 `c_char` is unsigned (`u8`). On x86_64 `c_char` is signed (`i8`) but on
aarch it's unsigned (`u8`). It was hard coded as `i8` so Rust refuses to compile it on aarch64.

Also make the type explicit in `broadcast_id` in case that causes issues
later, it should make them more obvious.